Refactor code to use the Parser class
Was previously dumping all logic of input and output in the UI class.

Extracted the parsing of inputs into the Parser class, formatting and passing the output back to the UI class, letting UI class do the exception handling and calling only.

This is the ideal way OOP should be done in this case, however we may still further SLAP the parser class.

* also renamed unmarkAsDone() to unmark() because it didn't make sense